types...

+++ src/interface.jl
@@ -2,11 +2,11 @@ using Dates
 using JSON3
 
 mutable struct LintResult
-    files_count::Int64
-    violations_count::Int64
-    recommendations_count::Int64
+    files_count::Integer
+    violations_count::Integer
+    recommendations_count::Integer
     linted_files::Vector{String}
-    printout_count::Int64
+    printout_count::Integer
 
     LintResult(a, b, c, d, e) = new(a, b, c, d, e)
 end
@@ -150,7 +150,7 @@ function convert_offset_to_line_from_filename(offset::Union{Int64, Int32}, filen
     return convert_offset_to_line_from_lines(offset, all_lines)
 end
 
-function convert_offset_to_line(offset::Int, source::String)
+function convert_offset_to_line(offset::Integer, source::String)
     return convert_offset_to_line_from_lines(offset, split(source, "\n"))
 end
 
@@ -367,8 +367,8 @@ end
 function print_summary(
     ::PlainFormat,
     io::IO,
-    count_violations::Int,
-    count_recommendations::Int
+    count_violations::Integer,
+    count_recommendations::Integer
 )
     nb_hints = count_violations + count_recommendations
     if iszero(nb_hints)
@@ -555,9 +555,9 @@ end
 function print_datadog_report(
     json_output::IO,
     report_as_string::String,
-    files_count::Int64,
-    violation_count::Int64,
-    recommandation_count::Int64,
+    files_count::Integer,
+    violation_count::Integer,
+    recommandation_count::Integer,
 )
     event = Dict(
         :source => "StaticLint",
